Fixed iGPU and dGPU. Updated kexts to macOS Ventura.
Added ACPI file to completely disable loading of the dGPU (before it was only soft-disabled).
Fixed config.plist configurations for the iGPU that were causing issues with the power management on battery.
Updated kexts that wouldn't work on macOS Ventura and up versions.
Fixed other issues related to the bluetooth on Monterey and up macOS versions.
Added HoRNDIS.kext to be able to use Android phones as USB hotspots.